Rename the app core folder
