+++ .htaccess
@@ -1,6 +1,6 @@
 RewriteEngine On
 
-RewriteBase /notes
+RewriteBase /tiny_mvc
 
 RewriteCond %{REQUEST_FILENAME} !-f
 RewriteCond %{REQUEST_FILENAME} !-d
