+++ App/Core/Enums/HttpMethod.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\AppCore\Enums;
+namespace App\Core\Enums;
 
 enum HttpMethod: string
 {
