+++ App/Core/Utils/Sanitize.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\AppCore\Utils;
+namespace App\Core\Utils;
 
 // Jednotuchá třída na základní ošetření vstupů
 class Sanitize
