+++ App/Core/Utils/Session.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\AppCore\Utils;
+namespace App\Core\Utils;
 
 /*
  * Pomocná třída pro dočasné ukládání dat to $_SESSION
